+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [@bpmn-io/feel-analyzer](https://github.com/bpmn-io/feel-
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: extract called function names from expressions ([#6](https://github.com/bpmn-io/feel-analyzer/pull/6))
+
 ## 0.1.0
 
 * `FEAT`: initial release, analyse inputs and their structure ([#4](https://github.com/bpmn-io/feel-analyzer/pull/4))